pyproject.toml
[project]
name = "ai-as-a-backend"
version = "0.1.0"
description = "AI-as-a-Backend research project using FunctionGemma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]
finetuning = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "trl>=0.8.0",
    "tensorboard>=2.16.0",
    "protobuf>=4.25.0",
    "sentencepiece>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
